final structure
