+++ requirements.txt
@@ -1,4 +1,7 @@
 torch
+torchvision
+skimage
+cv2
 cycler==0.10.0
 graphviz==0.10.1
 imbalanced-learn==0.4.3
