+++ src/cnn/SRM_filters.py
